+++ Public/Graph/New-MermaidGraphNodeConnection.ps1
@@ -1,4 +1,4 @@
-function New-MermaidNodeConnection {
+function New-MermaidGraphNodeConnection {
     [CmdletBinding()]
     param (
         $FirstNode,