+++ example.ps1
@@ -1,30 +1,30 @@
 New-MermaidGraph -Direction LR -NodeConnections @(
     $(
         $newMermaidNodeConnectionSplat = @{
-            FirstNode = $(New-MermaidNode -Shape RoundEdges -ID ID1 -Text "Frankfurt am Main" -Class "Starter")
-            SecondNode = $(New-MermaidNode -Shape DoubleCircle -ID ID2 -Text "PSConfEU$((Get-Date).Year +1)")
-            Link = $(New-MermaidLink -Text "traveling to:" -Linktype Link -ArrowType Cross) 
+            FirstNode = $(New-MermaidGraphNode -Shape RoundEdges -ID ID1 -Text "Frankfurt am Main" -Class "Starter")
+            SecondNode = $(New-MermaidGraphNode -Shape DoubleCircle -ID ID2 -Text "PSConfEU$((Get-Date).Year +1)")
+            Link = $(New-MermaidGraphLink -Text "traveling to:" -Linktype Link -ArrowType Cross) 
         }
-        New-MermaidNodeConnection @newMermaidNodeConnectionSplat
+        New-MermaidGraphNodeConnection @newMermaidNodeConnectionSplat
     ),
     $(
         $newMermaidNodeConnectionSplat = @{
-            FirstNode = $(New-MermaidNode -Shape DoubleCircle -ID ID2 -Text "PSConfEU$((Get-Date).Year +1)")
-            SecondNode = $(New-MermaidNode -Shape subroutine -ID ID3 -Text "Enjoying for 4 Days")
-            Link = $(New-MermaidLink -Linktype Dotted -ArrowType Arrow -BiDirectional)
+            FirstNode = $(New-MermaidGraphNode -Shape DoubleCircle -ID ID2 -Text "PSConfEU$((Get-Date).Year +1)")
+            SecondNode = $(New-MermaidGraphNode -Shape subroutine -ID ID3 -Text "Enjoying for 4 Days")
+            Link = $(New-MermaidGraphLink -Linktype Dotted -ArrowType Arrow -BiDirectional)
         }
-        New-MermaidNodeConnection @newMermaidNodeConnectionSplat
+        New-MermaidGraphNodeConnection @newMermaidNodeConnectionSplat
     ),
     $(
         $newMermaidNodeConnectionSplat = @{
-            FirstNode = $(New-MermaidNode -Shape DoubleCircle -ID ID2 -Text "PSConfEU$((Get-Date).Year +1)")
-            SecondNode = $(New-MermaidNode -Shape RoundEdges -ID ID1 -Text "Frankfurt am Main")
-            Link = $(New-MermaidLink -Text "traveling home:" -Linktype Link -ArrowType Dot)
+            FirstNode = $(New-MermaidGraphNode -Shape DoubleCircle -ID ID2 -Text "PSConfEU$((Get-Date).Year +1)")
+            SecondNode = $(New-MermaidGraphNode -Shape RoundEdges -ID ID1 -Text "Frankfurt am Main")
+            Link = $(New-MermaidGraphLink -Text "traveling home:" -Linktype Link -ArrowType Dot)
         }
-        New-MermaidNodeConnection @newMermaidNodeConnectionSplat
+        New-MermaidGraphNodeConnection @newMermaidNodeConnectionSplat
     )
 ) -ClassDefinitions @(
-    New-MermaidClassDefinition -Name "Starter" -FillColor "#6699ff" -StrokeColor "#999966"
+    New-MermaidGraphStyleClassDefinition -Name "Starter" -FillColor "#6699ff" -StrokeColor "#999966"
 )